+++ apps/test/server-mode/tailwind.config.mjs
@@ -1,0 +1,24 @@
+import path from 'node:path'
+
+/** @type {import('tailwindcss').Config} */
+export default {
+  content: [
+    './{blog,docs,src}/**/*.{astro,html,js,jsx,md,mdx,svelte,ts,tsx,vue}',
+    path.join(
+      path.dirname(require.resolve('@levino/shipyard-base')),
+      '../astro/**/*.astro',
+    ),
+    path.join(
+      path.dirname(require.resolve('@levino/shipyard-docs')),
+      '../astro/**/*.astro',
+    ),
+    path.join(
+      path.dirname(require.resolve('@levino/shipyard-blog')),
+      '../astro/**/*.astro',
+    ),
+  ],
+  theme: {
+    extend: {},
+  },
+  plugins: [require('@tailwindcss/typography'), require('daisyui')],
+}